backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-pptx==0.6.21
pydantic==2.3.0
python-multipart==0.0.6
watchfiles==0.20.0
pdf2image==1.17.0
openai==1.12.0
python-dotenv==1.0.0
tiktoken==0.5.2
# La librería UNO debe instalarse junto con LibreOffice
# No se puede instalar directamente con pip 
# FFmpeg debe estar instalado en el sistema (no es un paquete Python)
pillow
markdown
fpdf2
aiohttp
numpy
psutil>=7.0.0 